refacto(collection): refactor MyCollectionPage and UserProfilePage using widgets separate
